@@ -1,14 +1,13 @@
 name: listanything
 description: List anything, sometimes with a map
-version: 0.0.1+0000001
+version: 0.0.2+0000001
 publish_to: none
 
 environment:
-  sdk: ">=2.16.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
-  build_runner: ^2.3.0
-  build_verify: ^3.1.0
+  cloud_firestore: ^4.4.3
   device_info_plus: ^8.1.0
   firebase_auth: ^4.2.8
   firebase_core: ^2.6.1
@@ -18,21 +17,30 @@ dependencies:
   flutter:
     sdk: flutter
   flutter_bloc: ^8.0.1
+  flutter_form_builder: ^7.7.0
   flutter_hooks: ^0.18.6
   flutter_localizations:
     sdk: flutter
   flutter_riverpod: ^2.1.1
+  form_builder_validators: ^8.4.0
   go_router: ^6.0.9
-  go_router_builder: ^1.1.1
   intl: ^0.17.0
   riverpod: ^2.2.0
 
 dev_dependencies:
+  build_runner: ^2.3.0
+  build_verify: ^3.1.0
   flutter_test:
     sdk: flutter
+  freezed: ^2.2.0
+  go_router_builder: ^1.1.1
+  json_serializable: ^6.5.4
   mocktail: ^0.3.0
   very_good_analysis: ^2.4.0
 
 flutter:
   uses-material-design: true
   generate: true
+
+  assets:
+    - assets/images/
