@@ -4,13 +4,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:listanything/app/firebase/firebase_auth_provider.dart';
 import 'package:listanything/app/pages/about/about_page.dart';
-import 'package:listanything/app/pages/counter/counter_page.dart';
+import 'package:listanything/app/pages/lists/lists_page.dart';
 import 'package:listanything/app/widgets/standardWidgets/async_value_widget.dart';
 
 final selectedIndexProvider = StateProvider((ref) => 0);
 
 class HomeScreen extends ConsumerWidget {
-  const HomeScreen({Key? key}) : super(key: key);
+  const HomeScreen({super.key});
 
   static const TextStyle optionStyle = TextStyle(fontSize: 30, fontWeight: FontWeight.bold);
 
@@ -22,7 +22,7 @@ class HomeScreen extends ConsumerWidget {
       value: ref.watch(firebaseAuthProvider),
       data: (auth) {
         final _widgetOptions = <Widget>[
-          const CounterPage(),
+          const ListsPage(),
           ProfileScreen(auth: auth),
           const AboutPage(),
         ];
@@ -37,8 +37,8 @@ class HomeScreen extends ConsumerWidget {
           bottomNavigationBar: BottomNavigationBar(
             items: const <BottomNavigationBarItem>[
               BottomNavigationBarItem(
-                icon: Icon(Icons.plus_one),
-                label: 'Counter',
+                icon: Icon(Icons.account_tree_sharp),
+                label: 'Firestore',
               ),
               BottomNavigationBarItem(
                 icon: Icon(Icons.person),
