@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class AsyncValueWidget<T> extends StatelessWidget {
-  const AsyncValueWidget({Key? key, required this.value, required this.data}) : super(key: key);
+  const AsyncValueWidget({super.key, required this.value, required this.data});
   final AsyncValue<T> value;
   final Widget Function(T) data;
 
